경로 패치, 리뷰 더보기 기능 보완

+++ HarvestMovie/src/main/webapp/js/review.js
@@ -15,32 +15,36 @@ $(function () {
             cache: false,
             timeout: 30000,
             success: function (param) {
-                $(param.reviewVOList).each(function (index, item) {
-                    let output =
-                        '<div class="row d-flex justify-content-center mb-3">' +
-                        '<div class="card col col-sm-6 ">' +
-                        '<div class="card-body">' +
-                        '<h5 class="card-title">' + item.mem_name;
-                    if (item.isCritic == true) {
-                        output += '<img src="../../images/critic.svg" width="4%">';
-                    }
-                    output += '</h5>' +
-                        '<p class="card-text">' + item.review_message + '</p>' +
-                        '<div class="d-flex justify-content-end">';
-                    if (param.memberVO.auth == 9) {
-                        output += '<input type="button" value="삭제" class="btn btn-secondary' +
-                            'onClick="location.href=\'/movie/review/reviewDelete.do?mv_num=' +
-                            item.mv_num +
-                            '&rv_user_num=' + item.user_num + '\'">';
-                    }
+                if (param.reviewVOList.length > 0) {
+                    $(param.reviewVOList).each(function (index, item) {
+                        let output =
+                            '<div class="row d-flex justify-content-center mb-3">' +
+                            '<div class="card col col-sm-6 ">' +
+                            '<div class="card-body">' +
+                            '<h5 class="card-title">' + item.mem_name;
+                        if (item.isCritic == true) {
+                            output += '<img src="../../images/critic.svg" width="4%">';
+                        }
+                        output += '</h5>' +
+                            '<p class="card-text">' + item.review_message + '</p>' +
+                            '<div class="d-flex justify-content-end">';
+                        if (param.memberVO.auth == 9) {
+                            output += '<input type="button" value="삭제" class="btn btn-secondary' +
+                                'onClick="location.href=\'/movie/review/reviewDelete.do?mv_num=' +
+                                item.mv_num +
+                                '&rv_user_num=' + item.user_num + '\'">';
+                        }
 
-                    output += '</div>' +
-                        '</div>' +
-                        '</div>' +
-                        '</div>';
-                    $('#test').before(output);
+                        output += '</div>' +
+                            '</div>' +
+                            '</div>' +
+                            '</div>';
+                        $('#test').before(output);
 
-                })
+                    })
+                } else {
+                    alert('리뷰가 없습니다.')
+                }
 
 
             },
